requirements.txt
asgiref==3.6.0
Django==4.1.5
lxml==4.9.2
python-docx==0.8.11
sqlparse==0.4.3
